Created the dashboard_api app. It currently returns the total uplink and downlink (summary).
Create the core app. It includes the models for virtual system, rule dictionary and logs.

The DC engine has been modified to process csv in sequential order. The virtual_system column in the output csv has been renamed to virtual_system_id to account for the column created in db by django.

The DB engine has been modified to process csv in sequential order. The date column is added to the csv file and written to db. The date in the date column is calculated from the date in the name of the csv file.
Also, the username, password and db name have been hidden for security reasons, by exporting them in environment variables and using os.environ.get.